Selected text and gutter updates
Changed the background of selected text so it stands out from the
current line marker so you can see what is actually selected in the
line. Changed the text color of the line numbers and other text in the
gutter, as well as leaving the background color the same for the current
and selected lines.